New helper: query_params, returns a simplistic hash of parameters from the query string
Enhancements to ModelQueries: sql method for adding raw sql, block method for 'importing' another query block into this query, all? and any? for performing an 'and' or 'or' on all the query fragments passed.

Fixes to ModelQueries when the query includes joins

Various fixes to new http put/post mechanism, plus specs for the same

New data_filter mechanism for model controllers, replaces def_data_filter.

creator_attr mechanism on hobo models upgraded so it only sets the creator if the current user is of the correct type.

Small fixes in rapid pages

<TablePlus> now supports searching in addition to sorting

<hidden_fields for_query_string> will generate hidden fields for simple name/value pairs in the query string




git-svn-id: svn://hobocentral.net/hobo/branches/tom_sandbox@408 1a7aebd2-2222-0410-9f76-d2fcbb06e3f4