Changes to app-builder in hobo_spec
Renaming hoboPartPage to pagePath

Unused helper render_params removed

Hobo now tries to automatically re-render the correct action when a form submission has validation errors. Previously it was assumed that new or edit should be rendered. This is achieved with a hidden field added to every form.

The hobo command now uses Rails gems rather than svn checkout (not that Rails 1.99 is out)



git-svn-id: svn://hobocentral.net/hobo/branches/tom_sandbox@545 1a7aebd2-2222-0410-9f76-d2fcbb06e3f4