Scopes: new apply_scopes method for models and has_many associations
Pass a hash of scope-name => scope-arguments (single valur or array), the result will be scoped accordingly, but only if the first argument to the scope is not blank. It's designed to be used from controllers: if a param (e.g. a search or filter) is not included in the request, the scope is not applied