Scopes: new apply_scopes method for models and has_many associations
Pass a hash of scope-name => scope-arguments (single valur or array), the result will be scoped accordingly, but only if the first argument to the scope is not blank. It's designed to be used from controllers: if a param (e.g. a search or filter) is not included in the request, the scope is not applied


git-svn-id: svn://hobocentral.net/hobo/trunk@1133 1a7aebd2-2222-0410-9f76-d2fcbb06e3f4