*Breaking Change* EnumString: Change to automatically defined class methods on EnumString classes
It used to be that if you did EnumString.for(:foo, :baa), your class had methods foo, baa and the instance got methods foo? and baa?. There was a big problem with name clashes with other class methods. You now get no class methods, and the instance methods are called is_foo? and is_baa?.