Changed @dont_clear_key to an attr_accessor.
In order to prevent a "secure link" from being used twice,
the key_timestamp is reset on any state change (including
the one using the "secure link").  This change (hopefully)
provides a hook that can be called to preven the
key_timestamp from being reset and allowing the "secure link"
to be used more than once.