Add label elements to field-list
Previously Rapid did not rander <label> elements inside the forms. This
for example made it impossible to use the labels in cucumber features.
The implementation is not the cleanest around because of two goals:
first be as compatible as possible with the existing implementation and
secondly to avoid calling into Rails internals to figure out the id.
[#759 state:resolved]